requirements.txt
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
streamlit==1.30.0
streamlit_authenticator==0.3.1
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.12
